chore: rename llpkg-tool -> tools
