--- llpkg-tool/go.mod
@@ -1,3 +1,0 @@
-module github.com/goplus/llpkg/llpkg-tool
-
-go 1.21.13
