+++ tools/go.mod
@@ -1,0 +1,3 @@
+module github.com/goplus/llpkg/tools
+
+go 1.21.13
