+++ tools/internal/cmdbuilder/cmdbuilder.go
