+++ tools/internal/cmdbuilder/cmdbuilder_test.go
